requirements.txt
newsapi_python
newspaper3k
nltk
